backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
nltk==3.8.1
joblib==1.3.2
python-dotenv==1.0.0
alembic==1.13.1
